refactor: reorganized utils and imports
